adding proper documentation to the ScaffoldComps
