+++ app/src/main/java/com/example/clicker/presentation/home/HomeFragment.kt
@@ -107,7 +107,6 @@ class HomeFragment : Fragment() {
                         onNavigate = { dest -> findNavController().navigate(dest) },
                         addToLinks = { context.startActivity(domainIntent) },
                         quarterTotalScreenHeight,
-                        loadingPadding
 
 
                     )
